Created initial directory structure and files.
Created the initial directory structure and files to be used with Object
Oriented Programming with PHP for a user login/logout system. The user
will be able to register a new account, set a remember me when logging
in, update their profile, change their password, sign in and sign out.
Other classes include the ability to create cookies for user sessions,
perform database CRUD, sanitize and validate user input, escape output,
create and use tokens, redirect the user to a different page, have a
hash class to hash passwords and other pieces of data, and use a custom
404 error page.